- Adding solr facility in xml configuration - Adding keyword "score" for sorting - Better score function handling - Logger use now standard JVM logger

+++ src/java/com/jaeksoft/searchlib/function/expression/operator/DivideExpression.java
@@ -32,7 +32,7 @@ public class DivideExpression extends OperatorExpression {
 
 	@Override
 	public float newValue(float value1, float value2) {
-		return value1 / value2;
+		return (float) value1 / value2;
 	}
 
 	public String toString() {
